@@ -32,11 +32,11 @@
   "devDependencies": {
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
-    "@mui/icons-material": "^5.0.0",
-    "@mui/material": "^5.0.0",
+    "@mui/icons-material": "^5.6.0",
+    "@mui/material": "^5.6.0",
     "@types/node": "^17.0.22",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "@types/webpack": "^5.28.0",
     "@types/webpack-dev-server": "^4.3.0",
     "@typescript-eslint/eslint-plugin": "^5.16.0",
@@ -52,8 +52,8 @@
     "html-webpack-plugin": "^5.5.0",
     "prettier": "^2.2.1",
     "prettier-plugin-packagejson": "^2.2.8",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "ts-loader": "^9.2.6",
     "typedoc": "^0.22.15",
     "typedoc-cross-link-types": "^1.0.2",
@@ -66,8 +66,6 @@
   },
   "peerDependencies": {
     "@mui/icons-material": "^5.0.0",
-    "@mui/material": "^5.0.0",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0"
+    "@mui/material": "^5.0.0"
   }
 }
