@@ -1,5 +1,5 @@
 import React, { forwardRef, useCallback, useRef, useState } from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import _sampleData from "./example/db/sampleData";
 import TreeSelect, {
   FreeSoloNode,
@@ -353,4 +353,5 @@ const App = () => {
   );
 };
 
-ReactDOM.render(<App />, document.getElementById("root"));
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+createRoot(document.getElementById("root")!).render(<App />);
